+++ lib/framework/WWWData.h
@@ -2,7 +2,7 @@
 #include <Arduino.h>
 // index.html
 const uint8_t ESP_SVELTEKIT_DATA_0[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0x85,0x52,0xB1,0x72,0xDC,0x20,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0x85,0x52,0xB1,0x72,0xDC,0x20,
 	0x10,0xAD,0x75,0x5F,0x81,0xA9,0xEE,0x0A,0x41,0x93,0x22,0x63,0x83,0x9A,0xC4,0x9D,
 	0x3D,0xC9,0x4C,0xDC,0xB8,0xBA,0x41,0x62,0x4F,0x10,0x23,0x60,0x04,0xD2,0xF9,0xC6,
 	0xE3,0x3F,0xF0,0x0F,0xA4,0xCA,0x2F,0xE6,0x13,0x02,0x48,0x72,0x26,0x4D,0xD2,0xB0,
@@ -33,7 +33,7 @@ const uint8_t ESP_SVELTEKIT_DATA_0[] = {
 
 // favicon.png
 const uint8_t ESP_SVELTEKIT_DATA_1[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0x7D,0xBA,0x53,0x70,0x25,0x0C,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0x7D,0xBA,0x53,0x70,0x25,0x0C,
 	0x13,0x2D,0x1A,0xDB,0xD9,0xC9,0xC4,0xB6,0x6D,0xDB,0xCE,0x44,0x13,0xDB,0xB6,0x93,
 	0x89,0x6D,0xDB,0xB6,0x6D,0x73,0xC7,0xE6,0xC4,0xB6,0xEF,0xF7,0xDF,0x87,0x7B,0xAB,
 	0xCE,0xC3,0xE9,0xAA,0xB5,0x56,0x63,0x55,0xF5,0x4B,0x3F,0x76,0x84,0x92,0x82,0x24,
@@ -747,7 +747,7 @@ const uint8_t ESP_SVELTEKIT_DATA_1[] = {
 
 // manifest.json
 const uint8_t ESP_SVELTEKIT_DATA_2[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0x1D,0x8D,0x4D,0x0E,0xC2,0x20,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0x1D,0x8D,0x4D,0x0E,0xC2,0x20,
 	0x14,0x84,0xD7,0x70,0x0A,0xF2,0xD6,0x46,0x23,0xA9,0xFD,0x3B,0x83,0x9E,0xC0,0x18,
 	0x43,0x90,0x56,0x12,0x7C,0x34,0x40,0x0D,0xB5,0xE9,0xDD,0xE5,0x75,0x31,0x9B,0xF9,
 	0xBE,0xCC,0xAC,0x9C,0x01,0xAA,0x8F,0x81,0x5E,0xC0,0xCD,0x7B,0xBC,0xDA,0xF1,0x9D,
@@ -762,21 +762,21 @@ const uint8_t ESP_SVELTEKIT_DATA_2[] = {
 
 // _app/version.json
 const uint8_t ESP_SVELTEKIT_DATA_3[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0xAB,0x56,0x2A,0x4B,0x2D,0x2A,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0xAB,0x56,0x2A,0x4B,0x2D,0x2A,
 	0xCE,0xCC,0xCF,0x53,0xB2,0x52,0x32,0x34,0x37,0x37,0x32,0x35,0x35,0x34,0x36,0x32,
 	0x34,0xB5,0x30,0x52,0xAA,0x05,0x00,0xA0,0x60,0xDB,0xE8,0x1B,0x00,0x00,0x00,
 };
 
 // _app/env.js
 const uint8_t ESP_SVELTEKIT_DATA_4[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0x4B,0xAD,0x28,0xC8,0x2F,0x2A,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0x4B,0xAD,0x28,0xC8,0x2F,0x2A,
 	0x51,0x48,0xCE,0xCF,0x2B,0x2E,0x51,0x48,0xCD,0x2B,0xB3,0xAD,0xAE,0x05,0x00,0x2C,
 	0x1E,0xAB,0xEF,0x13,0x00,0x00,0x00,
 };
 
 // _app/immutable/bundle.js
 const uint8_t ESP_SVELTEKIT_DATA_5[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0xBC,0xBD,0x89,0x7A,0xDB,0x46,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0xBC,0xBD,0x89,0x7A,0xDB,0x46,
 	0xD2,0x28,0xFA,0x2A,0x22,0x7F,0x1D,0x1E,0x20,0x6A,0xD1,0x92,0xED,0x64,0x1C,0xD0,
 	0x08,0x8F,0x77,0x3B,0xF1,0x16,0xDB,0x49,0xEC,0xD0,0x1C,0x05,0x22,0x9B,0x24,0x2C,
 	0xB2,0x41,0x03,0x4D,0x2D,0xA6,0x78,0x9F,0xE5,0x3E,0xCB,0x7D,0xB2,0x5B,0x4B,0x6F,
@@ -13740,7 +13740,7 @@ const uint8_t ESP_SVELTEKIT_DATA_5[] = {
 
 // _app/immutable/assets/bundle.css
 const uint8_t ESP_SVELTEKIT_DATA_6[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0xED,0x7D,0xDB,0xAE,0xE3,0xCA,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0xED,0x7D,0xDB,0xAE,0xE3,0xCA,
 	0x72,0xD8,0x7B,0xBE,0x82,0x67,0x0E,0x16,0x66,0x71,0x0E,0xC9,0x11,0xA9,0xBB,0xE4,
 	0x19,0x6C,0x67,0xC3,0x46,0x0C,0xF8,0xD8,0x40,0xEC,0x04,0xB0,0x67,0x06,0x01,0x25,
 	0x51,0x12,0xCF,0x50,0xA2,0x4C,0x52,0xEB,0x32,0x8A,0xFC,0x9C,0xA7,0x3C,0x04,0x01,
@@ -14834,7 +14834,7 @@ const uint8_t ESP_SVELTEKIT_DATA_6[] = {
 
 // _app/immutable/assets/logo.png
 const uint8_t ESP_SVELTEKIT_DATA_7[] = {
-	0x1F,0x8B,0x08,0x00,0xEC,0x05,0xA7,0x69,0x02,0xFF,0x00,0x22,0x40,0xDD,0xBF,0x89,
+	0x1F,0x8B,0x08,0x00,0xA2,0x43,0xA7,0x69,0x02,0xFF,0x00,0x22,0x40,0xDD,0xBF,0x89,
 	0x50,0x4E,0x47,0x0D,0x0A,0x1A,0x0A,0x00,0x00,0x00,0x0D,0x49,0x48,0x44,0x52,0x00,
 	0x00,0x01,0x40,0x00,0x00,0x01,0x40,0x08,0x03,0x00,0x00,0x00,0xFA,0x4E,0x55,0x98,
 	0x00,0x00,0x03,0x00,0x50,0x4C,0x54,0x45,0x00,0x00,0x00,0x7D,0x75,0x8F,0xB9,0xC9,
