+++ platformio.ini
@@ -318,7 +318,7 @@ test_framework = doctest
 ; Override inherited [env] settings that are ESP32-specific
 framework =
 extra_scripts =
-lib_deps =
+lib_deps = ArduinoJson@>=7.0.0
 lib_ignore = PsychicHttp, framework
 build_src_filter = -<*> +<native_stub.cpp>
 lib_compat_mode = off
