+++ src/MoonBase/Nodes.cpp
@@ -12,7 +12,92 @@
 #if FT_MOONLIGHT
   #include "Nodes.h"
 
-  #include <ESP32SvelteKit.h>  //for safeModeMB
+String buildNameAndTags(const char* name, uint8_t dim, const char* tags) {
+  String result = name;
+
+  if (dim == _0D)
+    result += " 💡";
+  else if (dim == _1D)
+    result += " 📏";
+  else if (dim == _2D)
+    result += " ⏹️";
+  else if (dim == _3D)
+    result += " 🧊";
+
+  if (strlen(tags)) {
+    result += " ";
+    result += tags;
+  }
+
+  return result;
+}
+
+JsonObject Node::findOrCreateControl(const char* name, bool& newControl) {
+  JsonObject control = JsonObject();
+  for (JsonObject control1 : controls) {
+    if (control1["name"] == name) {
+      control = control1;
+      break;
+    }
+  }
+  if (control.isNull()) {
+    control = controls.add<JsonObject>();
+    control["name"] = name;
+    newControl = true;
+  }
+  return control;
+}
+
+JsonObject Node::setupControl(const char* name, const char* type, int min, int max, bool ro, const char* desc,
+                               uint8_t sizeCode, size_t sizeofVar, bool newControl, JsonObject control) {
+  control["type"] = type;
+  control["valid"] = true;
+  // optional properties
+  if (ro)
+    control["ro"] = true;
+  else if (!control["ro"].isNull())
+    control.remove("ro");
+  if (min != 0)
+    control["min"] = min;
+  else if (!control["min"].isNull())
+    control.remove("min");
+  if (max != UINT8_MAX)
+    control["max"] = max;
+  else if (!control["max"].isNull())
+    control.remove("max");
+  if (desc)
+    control["desc"] = desc;
+  else if (!control["desc"].isNull())
+    control.remove("desc");
+
+  // set size based on control type and sizeCode
+  if (control["type"] == "slider" || control["type"] == "select" || control["type"] == "pin" || control["type"] == "number") {
+    if (sizeCode == 8 || sizeCode == 108 || sizeCode == 16 || sizeCode == 32 || sizeCode == 33 || sizeCode == 34)
+      control["size"] = sizeCode;
+    else
+      EXT_LOGE(ML_TAG, "size %d mismatch for %s", sizeCode, name);
+  } else if (control["type"] == "selectFile" || control["type"] == "text") {
+    control["size"] = sizeofVar;
+  } else if (control["type"] == "checkbox") {
+    if (sizeCode != sizeof(bool))
+      EXT_LOGE(ML_TAG, "type for %s is not bool", name);
+    else
+      control["size"] = sizeof(bool);
+  } else if (control["type"] == "coord3D") {
+    if (sizeCode != sizeof(Coord3D))
+      EXT_LOGE(ML_TAG, "type for %s is not Coord3D", name);
+    else
+      control["size"] = sizeof(Coord3D);
+  } else
+    EXT_LOGE(ML_TAG, "type of %s not compatible: %s (%d)", control["name"].as<const char*>(), control["type"].as<const char*>(), control["size"].as<uint8_t>());
+
+  if (newControl) {
+    Char<20> oldValue = "";
+    onUpdate(oldValue, control);  // custom onUpdate for the node
+  }
+
+  return control;
+}
 
 void Node::updateControl(const JsonObject& control) {
   // EXT_LOGD(ML_TAG, "onUpdate %s", control["name"].as<const char*>());
@@ -72,530 +157,5 @@ void Node::updateControl(const JsonObject& control) {
     }
   }
 };
-
-  #if FT_LIVESCRIPT
-
-    #define USE_FASTLED  // as ESPLiveScript.h calls hsv ! one of the reserved functions!!
-    #include "ESPLiveScript.h"
-
-Node* gNode = nullptr;
-
-static void _addControl(uint8_t* var, char* name, char* type, uint8_t min = 0, uint8_t max = UINT8_MAX) {
-  EXT_LOGV(ML_TAG, "%s %s %d (%d-%d)", name, type, var, min, max);
-  gNode->addControl(*var, name, type, min, max);
-}
-static void _nextPin() { layerP.nextPin(); }
-static void _addLight(uint8_t x, uint8_t y, uint8_t z) { layerP.addLight({x, y, z}); }
-
-static void _modifySize() { gNode->modifySize(); }
-static void _modifyPosition(Coord3D& position) { gNode->modifyPosition(position); }  // need &position parameter
-// static void _modifyXYZ() {gNode->modifyXYZ();}//need &position parameter
-
-void _fadeToBlackBy(uint8_t fadeValue) { gNode->layer->fadeToBlackBy(fadeValue); }
-static void _setRGB(uint16_t indexV, CRGB color) { gNode->layer->setRGB(indexV, color); }
-static void _setRGBPal(uint16_t indexV, uint8_t index, uint8_t brightness) { gNode->layer->setRGB(indexV, ColorFromPalette(PartyColors_p, index, brightness)); }
-static void _setPan(uint16_t indexV, uint8_t value) { gNode->layer->setPan(indexV, value); }
-static void _setTilt(uint16_t indexV, uint8_t value) { gNode->layer->setTilt(indexV, value); }
-
-volatile xSemaphoreHandle WaitAnimationSync = xSemaphoreCreateBinary();
-
-void sync() {
-  static uint32_t frameCounter = 0;
-  frameCounter++;
-  delay(1);  // feed the watchdog, otherwise watchdog will reset the ESP
-  // Serial.print("s");
-  // 🌙 adding semaphore wait too long logging
-  if (xSemaphoreTake(WaitAnimationSync, pdMS_TO_TICKS(100)) == pdFALSE) {
-    EXT_LOGW(ML_TAG, "WaitAnimationSync wait too long");
-    xSemaphoreTake(WaitAnimationSync, portMAX_DELAY);
-  }
-}
-
-void addExternal(string definition, void* ptr) {
-  bool success = false;
-  size_t firstSpace = definition.find(' ');
-  if (firstSpace != std::string::npos) {
-    string returnType = definition.substr(0, firstSpace);
-
-    string parameters = "";
-    string functionName = "";
-
-    size_t openParen = definition.find('(', firstSpace + 1);
-    if (openParen != std::string::npos) {  // function
-      functionName = definition.substr(firstSpace + 1, openParen - firstSpace - 1);
-
-      size_t closeParen = definition.find(')', openParen + 1);
-      if (closeParen != std::string::npos) {  // function with parameters
-        parameters = definition.substr(openParen + 1, closeParen - openParen - 1);
-        success = true;
-        if (findLink(functionName, externalType::function) == -1)  // not allready added earlier
-          addExternalFunction(functionName, returnType, parameters, ptr);
-      }
-    } else {  // variable
-      functionName = definition.substr(firstSpace + 1);
-      success = true;
-      if (findLink(functionName, externalType::value) == -1)  // not allready added earlier
-        addExternalVariable(functionName, returnType, "", ptr);
-    }
-  }
-  if (!success) {
-    EXT_LOGE(ML_TAG, "Failed to parse function definition: %s", definition.c_str());
-  }
-}
-
-Parser parser = Parser();
-
-void LiveScriptNode::setup() {
-  // EXT_LOGV(ML_TAG, "animation %s", animation);
-
-  if (animation[0] != '/') {  // no sc script
-    return;
-  }
-
-  // make sure types in below functions are correct !!! otherwise livescript will crash
-
-  // generic functions
-  addExternal("uint32_t millis()", (void*)millis);
-  addExternal("uint32_t now()", (void*)millis);  // todo: synchronized time (sys->now)
-  addExternal("uint16_t random16(uint16_t)", (void*)(uint16_t (*)(uint16_t))random16);
-  addExternal("void delay(uint32_t)", (void*)((void (*)(uint32_t))delay));
-  addExternal("void pinMode(uint8_t,uint8_t)", (void*)pinMode);
-  addExternal("void digitalWrite(uint8_t,uint8_t)", (void*)digitalWrite);
-
-  // trigonometric functions
-  addExternal("float sin(float)", (void*)(float (*)(float))sin);
-  addExternal("float cos(float)", (void*)(float (*)(float))cos);
-  addExternal("uint8_t sin8(uint8_t)", (void*)sin8);
-  addExternal("uint8_t cos8(uint8_t)", (void*)cos8);
-  addExternal("float atan2(float,float)", (void*)(float (*)(float, float))atan2);
-  addExternal("uint8_t inoise8(uint16_t,uint16_t,uint16_t)", (void*)(uint8_t (*)(uint16_t, uint16_t, uint16_t))inoise8);
-  addExternal("uint8_t beatsin8(uint16_t,uint8_t,uint8_t,uint32_t,uint8_t)", (void*)beatsin8);
-  addExternal("float hypot(float,float)", (void*)(float (*)(float, float))hypot);
-  addExternal("float beat8(uint8_t,uint32_t)", (void*)beat8);  // saw wave
-  addExternal("uint8_t triangle8(uint8_t)", (void*)triangle8);
-
-  // MoonLight functions
-  addExternal("void addControl(void*,char*,char*,uint8_t,uint8_t)", (void*)_addControl);
-  addExternal("void nextPin()", (void*)_nextPin);
-  addExternal("void addLight(uint8_t,uint8_t,uint8_t)", (void*)_addLight);
-  addExternal("void modifySize()", (void*)_modifySize);
-  //   addExternal(    "void modifyPosition(Coord3D &position)", (void *)_modifyPosition);
-  //   addExternal(    "void modifyXYZ(uint16_t,uint16_t,uint16_t)", (void *)_modifyXYZ);
-
-  // MoonLight Parallel LED Driver vars
-  //   but keep enabled to avoid compile errors when used in non virtual context
-  //   addExternal( "uint8_t colorOrder", &layerP.ledsDriver.colorOrder);
-  //   addExternal( "uint8_t clockPin", &layerP.ledsDriver.clockPin);
-  //   addExternal( "uint8_t latchPin", &layerP.ledsDriver.latchPin);
-  //   addExternal( "uint8_t clockFreq", &layerP.ledsDriver.clockFreq);
-  //   addExternal( "uint8_t dmaBuffer", &layerP.ledsDriver.dmaBuffer);
-
-  addExternal("void fadeToBlackBy(uint8_t)", (void*)_fadeToBlackBy);
-  addExternal("CRGB* leds", (void*)(CRGB*)layerP.lights.channelsE);
-  addExternal("void setRGB(uint16_t,CRGB)", (void*)_setRGB);
-  addExternal("void setRGBPal(uint16_t,uint8_t,uint8_t)", (void*)_setRGBPal);
-  addExternal("void setPan(uint16_t,uint8_t)", (void*)_setPan);
-  addExternal("void setTilt(uint16_t,uint8_t)", (void*)_setTilt);
-  addExternal("uint8_t width", &layer->size.x);
-  addExternal("uint8_t height", &layer->size.y);
-  addExternal("uint8_t depth", &layer->size.z);
-  addExternal("bool on", &on);
-
-  //   for (asm_external el: external_links) {
-  //       EXT_LOGV(ML_TAG, "elink %s %s %d", el.shortname.c_str(), el.name.c_str(), el.type);
-  //   }
-
-  //   runningPrograms.setPrekill(layerP.ledsDriver.preKill, layerP.ledsDriver.postKill); //for clockless driver...
-  runningPrograms.setFunctionToSync(sync);
-
-  compileAndRun();
-}
-
-void LiveScriptNode::loop() {
-  // Serial.print("l");
-  xSemaphoreGive(WaitAnimationSync);
-}
-
-void LiveScriptNode::onLayout() {
-  if (hasOnLayout()) {
-    EXT_LOGV(ML_TAG, "%s", animation);
-    scriptRuntime.execute(animation, "onLayout");
-  }
-}
-
-LiveScriptNode::~LiveScriptNode() {
-  EXT_LOGV(ML_TAG, "%s", animation);
-  scriptRuntime.kill(animation);
-}
-
-// LiveScriptNode functions
-
-void LiveScriptNode::compileAndRun() {
-  // send UI spinner
-
-  // run the recompile not in httpd but in main loopTask (otherwise we run out of stack space)
-  //  runInAppTask.push_back([&, animation, type, error] {
-  EXT_LOGV(ML_TAG, "%s", animation);
-  File file = ESPFS.open(animation);
-  if (file) {
-    Char<32> pre;
-    pre.format("#define NUM_LEDS %d\n", layer->nrOfLights);
-
-    std::string scScript = pre.c_str();
-    scScript += file.readString().c_str();
-    file.close();
-
-    if (scScript.find("setup()") != std::string::npos) hasSetupFunction = true;
-    if (scScript.find("loop()") != std::string::npos) hasLoopFunction = true;
-    if (scScript.find("onLayout()") != std::string::npos) hasOnLayoutFunction = true;
-    if (scScript.find("modifyPosition(") != std::string::npos) hasModifyFunction = true;
-    //   if (scScript.find("modifyXYZ(") != std::string::npos) hasModifier = true;
-
-    // add main function
-    scScript += "void main(){";
-    if (hasSetupFunction) scScript += "setup();";
-    if (hasLoopFunction) scScript += "while(true){if(on){loop();sync();}else delay(1);}";  // loop must pauze when layout changes pass == 1! delay to avoid idle
-    scScript += "}";
-
-    EXT_LOGV(ML_TAG, "script \n%s", scScript.c_str());
-
-    // EXT_LOGV(ML_TAG, "parsing %s", scScript.c_str());
-
-    Executable executable = parser.parseScript(&scScript);  // note that this class will be deleted after the function call !!!
-    executable.name = animation;
-    EXT_LOGV(ML_TAG, "parsing %s done", animation);
-    scriptRuntime.addExe(executable);  // if already exists, delete it first
-    EXT_LOGV(ML_TAG, "addExe success %s", executable.exeExist ? "true" : "false");
-
-    gNode = this;  // todo: this is not working well with multiple scripts running!!!
-
-    if (executable.exeExist) {
-      execute();
-    } else
-      EXT_LOGV(ML_TAG, "error %s", executable.error.error_message.c_str());
-
-    // send error to client ... not working yet
-    //  error.set(executable.error.error_message); //String(executable.error.error_message.c_str());
-    //  _state.data["nodes"][2]["error"] = executable.error.error_message;
-  }
-  // });
-
-  // stop UI spinner
-}
-
-void LiveScriptNode::execute() {
-  if (safeModeMB) {
-    EXT_LOGW(ML_TAG, "Safe mode enabled, not executing script %s", animation);
-    return;
-  }
-  EXT_LOGV(ML_TAG, "%s", animation);
-
-  requestMappings();  // requestMapPhysical and requestMapVirtual will call the script onLayout function (check if this can be done in case the script also has loop running !)
-
-  if (hasLoopFunction) {
-    // setup : create controls
-    // executable.execute("setup");
-    // send controls to UI
-    // executable.executeAsTask("main"); //background task (async - vs sync)
-    EXT_LOGV(ML_TAG, "%s executeAsTask main", animation);
-    scriptRuntime.executeAsTask(animation, "main");  // background task (async - vs sync)
-    // assert failed: xEventGroupSync event_groups.c:228 (uxBitsToWaitFor != 0)
-  } else {
-    EXT_LOGV(ML_TAG, "%s execute main", animation);
-    scriptRuntime.execute(animation, "main");
-  }
-  EXT_LOGV(ML_TAG, "%s execute started", animation);
-}
-
-void LiveScriptNode::kill() {
-  EXT_LOGV(ML_TAG, "%s", animation);
-  scriptRuntime.kill(animation);
-}
-
-void LiveScriptNode::free() {
-  EXT_LOGV(ML_TAG, "%s", animation);
-  scriptRuntime.free(animation);
-}
-
-void LiveScriptNode::killAndDelete() {
-  EXT_LOGV(ML_TAG, "%s", animation);
-  scriptRuntime.kill(animation);
-  // scriptRuntime.free(animation);
-  scriptRuntime.deleteExe(animation);
-};
-
-void LiveScriptNode::getScriptsJson(JsonArray scripts) {
-  for (Executable& exec : scriptRuntime._scExecutables) {
-    exe_info exeInfo = scriptRuntime.getExecutableInfo(exec.name);
-    JsonObject object = scripts.add<JsonObject>();
-    object["name"] = exec.name;
-    object["isRunning"] = exec.isRunning();
-    object["isHalted"] = exec.isHalted;
-    object["exeExist"] = exec.exeExist;
-    object["handle"] = exec.__run_handle_index;
-    object["binary_size"] = exeInfo.binary_size;
-    object["data_size"] = exeInfo.data_size;
-    object["error"] = exec.error.error_message;
-    object["kill"] = 0;
-    object["free"] = 0;
-    object["delete"] = 0;
-    object["execute"] = 0;
-    // EXT_LOGV(ML_TAG, "scriptRuntime exec %s r:%d h:%d, e:%d h:%d b:%d + d:%d = %d", exec.name.c_str(), exec.isRunning(), exec.isHalted, exec.exeExist, exec.__run_handle_index, exeInfo.binary_size,
-    // exeInfo.data_size, exeInfo.total_size);
-  }
-}
-
-  #endif  // FT_LIVESCRIPT
-
-  #if HP_ALL_DRIVERS
-I2SClocklessLedDriver ledsDriver;
-  #endif
-
-void DriverNode::setup() {
-  addControl(layerP.lights.header.lightPreset, "lightPreset", "select");
-  addControlValue("RGB");
-  addControlValue("RBG");
-  addControlValue("GRB");  // default WS2812
-  addControlValue("GBR");
-  addControlValue("BRG");
-  addControlValue("BGR");
-  addControlValue("RGBW");                      // e.g. 4 channel par/dmx light
-  addControlValue("GRBW");                      // rgbw LED eg. sk6812
-  addControlValue("WRGB");                      // rgbw ws2814 LEDs
-  addControlValue("Curtain GRB6");              // some LED curtains
-  addControlValue("Curtain RGB2040");           // curtain RGB2040
-  addControlValue("Lightbar RGBWYP");           // 6 channel par/dmx light with UV etc
-  addControlValue("MH BeeEyes 150W-15");     // 15 channels moving head, see https://moonmodules.org/MoonLight/moonlight/drivers/#art-net
-  addControlValue("MH BeTopper 19x15W-32");  // 32 channels moving head
-  addControlValue("MH 19x15W-24");              // 24 channels moving heads
-}
-
-void DriverNode::loop() {
-  LightsHeader* header = &layerP.lights.header;
-
-  // use ledsDriver LUT for super efficient leds dimming 🔥 (used by rgbwBufferMapping)
-
-  uint8_t brightness = (header->offsetBrightness == UINT8_MAX) ? header->brightness : 255;  // set brightness to 255 if offsetBrightness is set (fixture will do its own brightness)
-
-  if (brightness != brightnessSaved || layerP.maxPower != maxPowerSaved) {
-    // Use FastLED for setMaxPowerInMilliWatts stuff, don't use if more then 8096 LEDs, decent power is assumed then! Also in case of Art-Net to HUB75 panels this calculation is not using the right mW per LED
-    uint8_t correctedBrightness = layerP.lights.header.nrOfLights > 8096 ? brightness : calculate_max_brightness_for_power_mW((CRGB*)&layerP.lights.channelsD, layerP.lights.header.nrOfLights, brightness, layerP.maxPower * 1000);  // calculate_max_brightness_for_power_mW supports max 65K LEDs
-    // EXT_LOGD(ML_TAG, "setBrightness b:%d + p:%d -> cb:%d", brightness, layerP.maxPower, correctedBrightness);
-    ledsDriver.setBrightness(correctedBrightness);
-    brightnessSaved = brightness;
-    maxPowerSaved = layerP.maxPower;
-  }
-
-  #if HP_ALL_DRIVERS
-  if (savedColorCorrection.red != layerP.lights.header.red || savedColorCorrection.green != layerP.lights.header.green || savedColorCorrection.blue != layerP.lights.header.blue) {
-    ledsDriver.setGamma(layerP.lights.header.red / 255.0, layerP.lights.header.blue / 255.0, layerP.lights.header.green / 255.0, 1.0);
-    // EXT_LOGD(ML_TAG, "setColorCorrection r:%d, g:%d, b:%d (%d %d %d)", layerP.lights.header.red, layerP.lights.header.green, layerP.lights.header.blue,
-    // savedColorCorrection.red, savedColorCorrection.green, savedColorCorrection.blue);
-    savedColorCorrection.red = layerP.lights.header.red;
-    savedColorCorrection.green = layerP.lights.header.green;
-    savedColorCorrection.blue = layerP.lights.header.blue;
-  }
-  #else  // ESP32_LEDSDRIVER
-  CRGB correction;
-  uint8_t white;
-  ledsDriver.getColorCorrection(correction.red, correction.green, correction.blue, white);
-  if (correction.red != header->red || correction.green != header->green || correction.blue != header->blue) {
-    EXT_LOGD(ML_TAG, "setColorCorrection r:%d, g:%d, b:%d (%d %d %d)", header->red, header->green, header->blue, correction.red, correction.green, correction.blue);
-    ledsDriver.setColorCorrection(header->red, header->green, header->blue);
-  }
-  #endif
-}
-
-void DriverNode::onUpdate(const Char<20>& oldValue, const JsonObject& control) {
-  LightsHeader* header = &layerP.lights.header;
-
-  // EXT_LOGD(ML_TAG, "%s: %s ", control["name"].as<const char*>(), control["value"].as<String>().c_str());
-
-  if (control["name"] == "lightPreset") {
-    uint8_t oldChannelsPerLight = header->channelsPerLight;
-
-    header->resetOffsets();
-
-    switch (header->lightPreset) {
-    case lightPreset_RGB:
-      header->channelsPerLight = 3;
-      header->offsetRed = 0;
-      header->offsetGreen = 1;
-      header->offsetBlue = 2;
-      break;
-    case lightPreset_RBG:
-      header->channelsPerLight = 3;
-      header->offsetRed = 0;
-      header->offsetGreen = 2;
-      header->offsetBlue = 1;
-      break;
-    case lightPreset_GRB:
-      header->channelsPerLight = 3;
-      header->offsetRed = 1;
-      header->offsetGreen = 0;
-      header->offsetBlue = 2;
-      break;
-    case lightPreset_GBR:
-      header->channelsPerLight = 3;
-      header->offsetRed = 2;
-      header->offsetGreen = 0;
-      header->offsetBlue = 1;
-      break;
-    case lightPreset_BRG:
-      header->channelsPerLight = 3;
-      header->offsetRed = 1;
-      header->offsetGreen = 2;
-      header->offsetBlue = 0;
-      break;
-    case lightPreset_BGR:
-      header->channelsPerLight = 3;
-      header->offsetRed = 2;
-      header->offsetGreen = 1;
-      header->offsetBlue = 0;
-      break;
-    case lightPreset_RGBW:
-      header->channelsPerLight = 4;
-      header->offsetRed = 0;
-      header->offsetGreen = 1;
-      header->offsetBlue = 2;
-      header->offsetWhite = 3;
-      break;
-    case lightPreset_GRBW:
-      header->channelsPerLight = 4;
-      header->offsetRed = 1;
-      header->offsetGreen = 0;
-      header->offsetBlue = 2;
-      header->offsetWhite = 3;
-      break;
-    case lightPreset_WRGB:
-      header->channelsPerLight = 4;
-      header->offsetRed = 1;
-      header->offsetGreen = 2;
-      header->offsetBlue = 3;
-      header->offsetWhite = 0;
-      break;
-    case lightPreset_GRB6:
-      header->channelsPerLight = 6;
-      header->offsetRed = 1;
-      header->offsetGreen = 0;
-      header->offsetBlue = 2;
-      break;
-    case lightPreset_RGB2040:
-      // RGB2040 uses standard RGB offsets but has special channel remapping
-      // for dual-channel-group architecture (handled in VirtualLayer)
-      header->channelsPerLight = 3;
-      header->offsetRed = 0;
-      header->offsetGreen = 1;
-      header->offsetBlue = 2;
-      break;
-    case lightPreset_RGBWYP:
-      header->channelsPerLight = 6;
-      header->offsetRed = 0;
-      header->offsetGreen = 1;
-      header->offsetBlue = 2;
-      header->offsetWhite = 3;
-      break;
-    case lightPreset_MHBeeEyes150W15:
-      header->channelsPerLight = 15;  // set channels per light to 15 (RGB + Pan + Tilt + Zoom + Brightness)
-      header->offsetRGBW = 10;        // set offset for RGB lights in DMX map
-      header->offsetRed = 0;
-      header->offsetGreen = 1;
-      header->offsetBlue = 2;
-      header->offsetPan = 0;
-      header->offsetTilt = 1;
-      header->offsetZoom = 7;
-      header->offsetBrightness = 8;   // set offset for brightness
-      header->offsetGobo = 5;         // set offset for color wheel in DMX map
-      header->offsetBrightness2 = 3;  // set offset for color wheel brightness in DMX map    } //BGR
-      break;
-    case lightPreset_MHBeTopper19x15W32:
-      header->channelsPerLight = 32;
-      header->offsetRGBW = 9;
-      header->offsetRed = 0;
-      header->offsetGreen = 1;
-      header->offsetBlue = 2;
-      header->offsetRGBW1 = 13;
-      header->offsetRGBW2 = 17;
-      header->offsetRGBW3 = 24;
-      header->offsetPan = 0;
-      header->offsetTilt = 2;
-      header->offsetZoom = 5;
-      header->offsetBrightness = 6;
-      break;
-    case lightPreset_MH19x15W24:
-      header->channelsPerLight = 24;
-      header->offsetRGBW = 4;
-      header->offsetRed = 0;
-      header->offsetGreen = 1;
-      header->offsetBlue = 2;
-      header->offsetWhite = 3;
-      header->offsetPan = 0;
-      header->offsetTilt = 1;
-      header->offsetBrightness = 3;
-      header->offsetRGBW1 = 8;
-      header->offsetRGBW2 = 12;
-      header->offsetZoom = 17;
-      break;
-    default:
-      EXT_LOGW(ML_TAG, "Invalid lightPreset value: %d", header->lightPreset);
-      // Fall back to GRB (most common default)
-      header->lightPreset = lightPreset_GRB;
-      header->channelsPerLight = 3;
-      header->offsetRed = 1;
-      header->offsetGreen = 0;
-      header->offsetBlue = 2;
-      break;
-    }
-
-    EXT_LOGD(ML_TAG, "setLightPreset %d (cPL:%d, o:%d,%d,%d,%d)", header->lightPreset, header->channelsPerLight, header->offsetRed, header->offsetGreen, header->offsetBlue, header->offsetWhite);
-
-    // FASTLED_ASSERT(true, "oki");
-
-  #if HP_ALL_DRIVERS
-    #ifndef CONFIG_IDF_TARGET_ESP32P4
-    if (initDone) {
-      // ledsDriver.setOffsets(header->offsetRed, header->offsetGreen, header->offsetBlue, header->offsetWhite);
-
-      // if (oldChannelsPerLight != header->channelsPerLight)
-      //   restartNeeded = true; //in case
-    }
-    #endif
-
-  #else  // ESP32_LEDSDRIVER
-    if (ledsDriver.initLedsDone) {
-      ledsDriver.setOffsets(header->offsetRed, header->offsetGreen, header->offsetBlue, header->offsetWhite);
-
-      if (oldChannelsPerLight != header->channelsPerLight) restartNeeded = true;  // in case
-    }
-  #endif
-
-    lightPresetSaved = true;
-  }
-}
-
-void DriverNode::rgbwBufferMapping(uint8_t* packetRGBChannel, uint8_t* lightsRGBChannel) {
-  // use ledsDriver.__rbg_map[0]; for super fast brightness and gamma correction! see secondPixel in ESP32-LedDriver!
-  // apply the LUT to the RGB channels !
-
-  uint8_t red = lightsRGBChannel[0];
-  uint8_t green = lightsRGBChannel[1];
-  uint8_t blue = lightsRGBChannel[2];
-  // extract White from RGB
-  if (layerP.lights.header.offsetWhite != UINT8_MAX) {
-    // if white is filled, use that and do not extract rgbw
-    uint8_t white = lightsRGBChannel[3];
-    if (!white) {
-      white = MIN(MIN(red, green), blue);
-      red -= white;
-      green -= white;
-      blue -= white;
-    }
-    packetRGBChannel[layerP.lights.header.offsetWhite] = ledsDriver.__white_map[white];
-  }
-
-  packetRGBChannel[layerP.lights.header.offsetRed] = ledsDriver.__red_map[red];
-  packetRGBChannel[layerP.lights.header.offsetGreen] = ledsDriver.__green_map[green];
-  packetRGBChannel[layerP.lights.header.offsetBlue] = ledsDriver.__blue_map[blue];
-}
 
 #endif  // FT_MOONLIGHT
